fix: removed axios call for portal generation and used sdk instead

+++ src/controllers/portal/generate.ts
@@ -1,63 +1,40 @@
 import cli from "cli-ux";
 import * as fs from "fs-extra";
-import * as FormData from "form-data";
-
-import { baseURL } from "../../config/env";
-import { deleteFile, extractZipFile, unzipFile } from "../../utils/utils";
+import { deleteFile, extractZipFile, writeFileUsingReadableStream } from "../../utils/utils";
 import { GeneratePortalParams } from "../../types/portal/generate";
-import { AuthInfo, getAuthInfo } from "../../client-utils/auth-manager";
-
-import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
-
-import axiosInstance from "../../config/axios-config";
-
-// TODO: Remove after SDK is patched
-const downloadPortalAxios = async (zippedBuildFilePath: string, overrideAuthKey: string | null, configDir: string) => {
-  const formData = new FormData();
-  const authInfo: AuthInfo | null = await getAuthInfo(configDir);
-  formData.append("file", fs.createReadStream(zippedBuildFilePath));
-  const config: AxiosRequestConfig = {
-    headers: {
-      Authorization: authInfo || overrideAuthKey ? `X-Auth-Key ${authInfo?.authKey.trim() || overrideAuthKey}` : "",
-      ...formData.getHeaders()
-    },
-    responseType: "arraybuffer"
-  };
-  const { data }: AxiosResponse = await axiosInstance.post(`${baseURL}/portal`, formData, config);
-  return data;
-};
+import { ApiResponse, ContentType, FileWrapper } from "@apimatic/sdk";
 
 // Download Docs Portal
 export const downloadDocsPortal = async (
-  { zippedBuildFilePath, portalFolderPath, zippedPortalPath, overrideAuthKey, zip }: GeneratePortalParams,
-  configDir: string
-) => {
+  { zippedBuildFilePath, portalFolderPath, zippedPortalPath, docsPortalController, zip }: GeneratePortalParams) => {
   cli.action.start("Downloading portal");
 
   // Check if the build file exists for the user or not
   if (!(await fs.pathExists(zippedBuildFilePath))) {
     throw new Error("Build file doesn't exist");
   }
-  // TODO: ***CRITICAL*** Remove this call once the SDK is patched
-  const data: ArrayBuffer = await downloadPortalAxios(zippedBuildFilePath, overrideAuthKey, configDir);
 
-  await deleteFile(zippedBuildFilePath);
-  await fs.writeFile(zippedPortalPath, data);
+  const file: FileWrapper = new FileWrapper(fs.createReadStream(zippedBuildFilePath));
 
-  // TODO: Uncomment this code block when the SDK is patched
-  // const file: FileWrapper = new FileWrapper(fs.createReadStream(zippedBuildFilePath));
-  // const { result }: ApiResponse<NodeJS.ReadableStream | Blob> =
-  //   await docsPortalController.generateOnPremPortalViaBuildInput(file);
-  // if ((data as NodeJS.ReadableStream).readable) {
-  //   await writeFileUsingReadableStream(data as NodeJS.ReadableStream, zippedPortalPath);
-  if (!zip) {
-    await extractZipFile(zippedPortalPath, portalFolderPath);
-    await deleteFile(zippedPortalPath);
+  try {
+    const { result }: ApiResponse<NodeJS.ReadableStream | Blob> =
+    await docsPortalController.generateOnPremPortalViaBuildInput(ContentType.EnumMultipartformdata, file);
+
+    if ((result as NodeJS.ReadableStream).readable) {
+      await writeFileUsingReadableStream(result as NodeJS.ReadableStream, zippedPortalPath);
+      
+      if (!zip) {
+        await extractZipFile(zippedPortalPath, portalFolderPath);
+        await deleteFile(zippedPortalPath);
+      }
+      
+      cli.action.stop();
+      return zip ? zippedPortalPath : portalFolderPath;
+    } else {
+      throw new Error("Couldn't download the portal");
+    }
   }
-
-  cli.action.stop();
-  return zip ? zippedPortalPath : portalFolderPath;
-  // } else {
-  //   throw new Error("Couldn't download the portal");
-  // }
+  catch (err) {
+    throw new Error("Couldn't download the portal because: " + err);  
+  }
 };
